Recap-471 : Made changes to search request UI. Updated request item entity.
